requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
requests>=2.31.0
openai>=1.10.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.8.0
